requirements.txt
fastapi==0.104.0
uvicorn==0.23.2
pydantic==2.4.2
google-cloud-bigquery==3.11.4
python-dotenv==1.0.0
requests==2.31.0 